+++ indispensable-josef/src/commonMain/kotlin/at/asitplus/signum/indispensable/josef/KeyAttestationJwt.kt
@@ -86,6 +86,7 @@ data class KeyAttestationJwt(
     val status: JsonObject? = null,
 ) {
 
+    @Deprecated("To be removed in next release")
     fun serialize() = joseCompliantSerializer.encodeToString(this)
 
     override fun equals(other: Any?): Boolean {
@@ -127,6 +128,7 @@ data class KeyAttestationJwt(
     }
 
     companion object {
+        @Deprecated("To be removed in next release")
         fun deserialize(it: String) = catching {
             joseCompliantSerializer.decodeFromString<KeyAttestationJwt>(it)
         }